tic_stackexchange_processing/requirements.txt
# StackExchange Processing requirements
argparse
boto3
fcntl
ftfy
joblib
numpy
pandas
psutil
pyunpack
ray
tqdm
transformers>=4.0.0
pyyaml
